small modifications
- Change __str__ property of graphs
- Remove comments

+++ eagerx/src/eagerx/converters/space_ros_converters.py
@@ -98,7 +98,7 @@ class Space_Float32(SpaceConverter):
 
     @staticmethod
     @register.spec('Space_Float32', SpaceConverter)
-    def spec(spec: ConverterSpec, low=None, high=None, dtype='uint64'):
+    def spec(spec: ConverterSpec, low=None, high=None, dtype='float32'):
         params = dict(low=low, high=high, dtype=dtype)
         spec.set_parameters(params)
